ethers, eth-sig-util 전용 script 생성 & test에서 해당항목 제거
