+++ projects/eip-1271/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@metamask/eth-sig-util": "^8.2.0",
-    "ethereumjs-util": "^7.1.5"
+    "ethereumjs-util": "^7.1.5",
+    "ethers": "6.13.5"
   }
 }
